completed all the functionalities of the website

+++ frontend/src/components/BlogDetails/BlogDetails.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { useHistory, useParams } from 'react-router-dom/cjs/react-router-dom.min'
 import './blogdetails.css'
 import useFetch from '../../hooks/useFetch'
@@ -8,29 +8,36 @@ const BlogDetails = () => {
   const history = useHistory()
   const { id } = useParams()
   const {user} = useAuthContext()
+  const allowdeleteBtn = localStorage.getItem('del')
 
-  const { data: blogs, isPending } = useFetch(`http://localhost:3001/api/blogs/${id}`)
+  const { data: blog, isPending } = useFetch(`http://localhost:3001/api/blogs/${id}`)
+
+  console.log(blog)
 
   const handleDelete = () => {
+    if (!user) return
+
     fetch(`http://localhost:3001/api/blogs/${id}`, {
       method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     }).then(() => {
       history.push('/')
-    
     })
   }
 
   return (
     <div>
-      { blogs && 
+      { blog && 
           <div className='blog-details'>
           <div className="blog-about">
-            <p><h1 className='blog-title'>{blogs.title}</h1> by {blogs.author}</p>
+            <p><h1 className='blog-title'>{blog.title}</h1> by {blog.author}</p>
           </div>
           <div className="blog-body">
-            <p>{blogs.body}</p>
+            <p>{blog.body}</p>
           </div>
-          <button className='btn-delete-blog' onClick={handleDelete}>delete blog</button>
+          { allowdeleteBtn && <button disabled={isPending} className='btn-delete-blog' onClick={handleDelete}>delete blog</button> }
         </div>
         }
     </div>
